Move encoding utilities into subfolder
